requirements.txt
# FastAPI & Server
fastapi==0.115.2
uvicorn[standard]==0.30.6

# Machine Learning / Embeddings
sentence-transformers==3.2.1
torch==2.4.1
transformers==4.45.2
Pillow==10.4.0

# Pinecone client (official new SDK)
pinecone==5.0.1

# Google Gemini (Text Embedding)
google-generativeai==0.7.2

# Utils
python-dotenv==1.0.1
requests==2.32.3